First attempt at making IChangeContext interface
